test_app/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.0
Django==4.2.6
django-cors-headers==4.3.0
djangorestframework==3.14.0
idna==3.4
psycopg2-binary==2.9.9
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.7
